Revert "feat: add getRenderHandler & optimize types (#5396)"
This reverts commit ac9cd104777d211d3d12c7acc9ad0122b332cfd4.
