+++ packages/server/core/src/base/adapters/env.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 import { fs, dotenv, dotenvExpand } from '@modern-js/utils';
-import { ServerBaseOptions } from '../../../core/server';
+import { ServerBaseOptions } from '../types';
 
 export async function loadServerEnv(options: ServerBaseOptions) {
   const { pwd } = options;
