+++ packages/server/core/src/base/adapters/monitor.ts
@@ -1,9 +1,8 @@
-import type { Next } from 'hono';
+import { Next } from 'hono';
 import { Logger, Reporter } from '@modern-js/types';
 import { time } from '@modern-js/runtime-utils/time';
+import { HonoContext, HonoNodeEnv } from '../types';
 import { ServerReportTimings } from '../libs/constants';
-import { HonoContext } from '../../core/server';
-import { HonoNodeEnv } from '../adapters/node';
 
 const defaultReporter: Reporter = {
   init() {
