+++ packages/server/core/src/base/adapters/stream.ts
