+++ packages/server/core/src/base/libs/customServer.ts
@@ -7,8 +7,7 @@ import {
   AfterStreamingRenderContext,
   ServerRoute,
 } from '@modern-js/types';
-import { HonoContext } from '../../core/server';
-import { HonoNodeEnv } from '../adapters/node';
+import { HonoContext, HonoNodeEnv } from '../types';
 import { createBaseHookContext } from './hook/base';
 import { RouterAPI } from './hook/routerApi';
 import { TemplateApi } from './hook/template';
