+++ packages/server/core/src/base/middlewares/customServer.ts
@@ -1,7 +1,7 @@
 import { ServerRoute } from '@modern-js/types';
 import { ServerHookRunner } from '@core/plugin';
 import { time } from '@modern-js/runtime-utils/time';
-import { Metrics, Middleware } from '../../core/server';
+import { Metrics, Middleware, HonoNodeEnv } from '../types';
 import {
   createAfterMatchCtx,
   createAfterRenderCtx,
@@ -11,7 +11,6 @@ import {
 import { createTransformStream } from '../libs/utils';
 import { ServerBase } from '../serverBase';
 import { ServerReportTimings } from '../libs/constants';
-import { HonoNodeEnv } from '../adapters/node';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-function
 const noop = () => {};
