+++ packages/server/core/src/base/renderHandler/index.ts
@@ -2,13 +2,12 @@ import path from 'path';
 import { existsSync } from 'fs';
 import { SERVER_DIR } from '@modern-js/utils';
 import { Metrics } from '@modern-js/types';
-import { Middleware, ServerBaseOptions } from '../../core/server';
+import { Middleware, ServerBaseOptions, HonoNodeEnv } from '../types';
 import { ServerBase } from '../serverBase';
 import { CustomServer } from '../middlewares';
 import { warmup } from '../libs/warmup';
 import { checkIsProd, getRuntimeEnv } from '../libs/utils';
-import { HonoNodeEnv } from '../adapters/node';
-import { initReporter } from '../middlewares/monitor';
+import { initReporter } from '../adapters/monitor';
 import { ssrCache } from './ssrCache';
 import { Render, createRender } from './render';
 
@@ -29,34 +28,17 @@ export type BindRenderHandleOptions = {
   staticGenerate?: boolean;
 };
 
-export function getRenderHandler(
-  options: ServerBaseOptions & BindRenderHandleOptions,
-) {
-  const { routes, pwd, config } = options;
-  if (routes && routes.length > 0) {
-    const ssrConfig = config.server?.ssr;
-    const forceCSR = typeof ssrConfig === 'object' ? ssrConfig.forceCSR : false;
-    const render = createRender({
-      routes,
-      pwd,
-      staticGenerate: options.staticGenerate,
-      metaName: options.metaName || 'modern-js',
-      forceCSR,
-      nonce: options.config.security?.nonce,
-    });
-    return render;
-  }
-  return null;
-}
-
 export async function bindRenderHandler(
   server: ServerBase,
   options: ServerBaseOptions & BindRenderHandleOptions,
 ) {
-  const { routes, pwd, metrics } = options;
+  const { config, routes, pwd, metrics } = options;
 
   const { runner } = server;
   if (routes && routes.length > 0) {
+    // TODO: get server config from server.ssr & server.ssrByEntries
+    const ssrConfig = config.server?.ssr;
+    const forceCSR = typeof ssrConfig === 'object' ? ssrConfig.forceCSR : false;
     const customServer = new CustomServer(runner, server, pwd, metrics);
 
     // warn ssr bundles
@@ -101,8 +83,15 @@ export async function bindRenderHandler(
       }
     }
 
-    const render = getRenderHandler(options);
+    const render = createRender({
+      routes,
+      pwd,
+      staticGenerate: options.staticGenerate,
+      metaName: options.metaName || 'modern-js',
+      forceCSR,
+      nonce: options.config.security?.nonce,
+    });
 
-    render && server.get('*', createRenderHandler(render));
+    server.get('*', createRenderHandler(render));
   }
 }
