--- packages/server/core/src/core/hono.ts
@@ -1,8 +1,0 @@
-export type {
-  NotFoundHandler,
-  Context as HonoContext,
-  Env as HonoEnv,
-  HonoRequest,
-  Next,
-  MiddlewareHandler as HonoMiddleware,
-} from 'hono';
