+++ packages/server/core/src/core/plugin.ts
@@ -1,3 +1,4 @@
+import { Readable } from 'stream';
 import type { IncomingMessage, ServerResponse } from 'node:http';
 import type { Component } from 'react';
 import {
@@ -26,8 +27,9 @@ import type {
   AfterStreamingRenderContext,
 } from '@modern-js/types';
 
+// TODO: should not import hono
+import type { MiddlewareHandler } from 'hono';
 import type { BffUserConfig, ServerOptions, UserConfig } from '../types/config';
-import { HonoMiddleware } from './hono';
 
 // collect all middleware register in server plugins
 const gather = createParallelWorkflow<{
@@ -77,7 +79,11 @@ export type APIServerStartInput = {
   config?: {
     middleware?: Array<any>;
   };
-  render?: HonoMiddleware | null;
+  render?: (
+    req: IncomingMessage,
+    res: ServerResponse,
+    url?: string,
+  ) => Promise<string | Readable | null>;
 };
 
 type Change = {
@@ -87,7 +93,7 @@ type Change = {
 
 const prepareApiServer = createAsyncPipeline<
   APIServerStartInput,
-  HonoMiddleware
+  MiddlewareHandler
 >();
 
 const onApiChange = createAsyncWaterfall<Change[]>();
